chore(channelPage): 컴포넌트 이름 수정 - 컴포넌트 server / client 구분
