+++ src/app/(route)/channel/[uid]/components/BoardInput.client.tsx
